Chatbot_bot/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
typing-extensions>=4.9.0
python-multipart>=0.0.9
httpx>=0.26.0
requests>=2.31.0